Refactor hyperopt parameters to separate file
