+++ freqtrade/strategy/__init__.py
@@ -1,9 +1,9 @@
 # flake8: noqa: F401
 from freqtrade.exchange import (timeframe_to_minutes, timeframe_to_msecs, timeframe_to_next_date,
                                 timeframe_to_prev_date, timeframe_to_seconds)
-from freqtrade.strategy.hyper import (BooleanParameter, CategoricalParameter, DecimalParameter,
-                                      IntParameter, RealParameter)
 from freqtrade.strategy.informative_decorator import informative
 from freqtrade.strategy.interface import IStrategy
+from freqtrade.strategy.parameters import (BooleanParameter, CategoricalParameter, DecimalParameter,
+                                           IntParameter, RealParameter)
 from freqtrade.strategy.strategy_helper import (merge_informative_pair, stoploss_from_absolute,
                                                 stoploss_from_open)
